Migrate logic to base
